Extracts video player from ImagePreviewViewController() and fixes use of deprecated method (UIApplication.shared.keyWindow)
