+++ piwigo/Image/ImagePreviewViewController.swift
@@ -29,7 +29,6 @@ class ImagePreviewViewController: UIViewController
     @IBOutlet weak var imageViewWidthConstraint: NSLayoutConstraint!
     @IBOutlet weak var imageViewHeightConstraint: NSLayoutConstraint!
     @IBOutlet weak var playImage: UIImageView!
-    @IBOutlet weak var videoView: UIView!
     @IBOutlet weak var descContainer: ImageDescriptionView!
     
     private var downloadTask: URLSessionDataTask?
@@ -363,118 +362,6 @@ class ImagePreviewViewController: UIViewController
                 self.view.layoutIfNeeded()
             }
         }
-    }
-        
-    
-    // MARK: - Video Player
-    @objc
-    func startVideoPlayerView(with imageData: PiwigoImageData?) {
-        // Set URL
-        let videoURL = URL(string: imageData?.fullResPath ?? "")
-
-        // AVURLAsset + Loader
-        var asset: AVURLAsset? = nil
-        if let videoURL = videoURL {
-            asset = AVURLAsset(url: videoURL, options: nil)
-        }
-        let loader = asset?.resourceLoader
-        loader?.setDelegate(self, queue: DispatchQueue(label: "Piwigo loader"))
-
-        // Load the asset's "playable" key
-        asset?.loadValuesAsynchronously(forKeys: ["playable"], completionHandler: { [self] in
-            DispatchQueue.main.async(
-                execute: { [self] in
-                    // IMPORTANT: Must dispatch to main queue in order to operate on the AVPlayer and AVPlayerItem.
-                    var error: NSError? = nil
-                    let keyStatus = asset?.statusOfValue(forKey: "playable", error: &error)
-                    switch keyStatus {
-                        case .loaded:
-                            // Sucessfully loaded, continue processing
-                            playVideoAsset(asset)
-                        case .failed:
-                            // Display the error.
-                            assetFailedToPrepare(forPlayback: error)
-                        case .cancelled:
-                            // Loading cancelled
-                            break
-                        default:
-                            // Handle all other cases
-                            break
-                    }
-                })
-        })
-    }
-
-    func playVideoAsset(_ asset: AVAsset?) {
-        // AVPlayer
-        var playerItem: AVPlayerItem? = nil
-        if let asset = asset {
-            playerItem = AVPlayerItem(asset: asset)
-        }
-        let videoPlayer = AVPlayer(playerItem: playerItem) // Intialise video controller
-        let playerController = AVPlayerViewController()
-        playerController.player = videoPlayer
-        playerController.videoGravity = .resizeAspect
-
-        // Playback controls
-        playerController.showsPlaybackControls = true
-//    [self.videoPlayer addObserver:self.imageView forKeyPath:@"rate" options:0 context:nil];
-
-        // Start playing automatically…
-        playerController.player?.play()
-
-        videoView?.addSubview(playerController.view)
-        playerController.view.frame = videoView?.bounds ?? CGRect.zero
-
-        // Present the video
-        playerController.modalTransitionStyle = .crossDissolve
-        playerController.modalPresentationStyle = .overCurrentContext
-        let currentViewController = UIApplication.shared.keyWindow?.topMostViewController()
-        currentViewController?.present(playerController, animated: true)
-    }
-
-    func assetFailedToPrepare(forPlayback error: Error?) {
-        // Determine the present view controller
-        if let error = error as NSError?,
-           let topViewController = UIApplication.shared.keyWindow?.topMostViewController() {
-            topViewController.dismissPiwigoError(withTitle: error.localizedDescription, message: "",
-                                                 errorMessage: error.localizedFailureReason ?? "",
-                                                 completion: { })
-        }
-    }
-}
-
-
-// MARK: - AVAssetResourceLoader Methods
-extension ImagePreviewViewController: AVAssetResourceLoaderDelegate
-{
-    func resourceLoader(_ resourceLoader: AVAssetResourceLoader,
-                        shouldWaitForResponseTo authenticationChallenge: URLAuthenticationChallenge
-    ) -> Bool {
-        let protectionSpace = authenticationChallenge.protectionSpace
-        if protectionSpace.authenticationMethod == NSURLAuthenticationMethodServerTrust
-        {
-            // Self-signed certificate…
-            if let certificate = protectionSpace.serverTrust {
-                let credential = URLCredential(trust: certificate)
-                authenticationChallenge.sender?.use(credential, for: authenticationChallenge)
-            }
-            authenticationChallenge.sender?.continueWithoutCredential(for: authenticationChallenge)
-        }
-        else if protectionSpace.authenticationMethod == NSURLAuthenticationMethodHTTPBasic {
-            // HTTP basic authentification credentials
-            let user = NetworkVars.httpUsername
-            let password = KeychainUtilities.password(forService: "\(NetworkVars.serverProtocol)\(NetworkVars.serverPath)", account: user)
-            authenticationChallenge.sender?.use(
-                URLCredential(user: user, password: password, persistence: .synchronizable),
-                              for: authenticationChallenge)
-            authenticationChallenge.sender?.continueWithoutCredential(for: authenticationChallenge)
-        }
-        else {
-            // Other type: username password, client trust...
-            print("Other type: username password, client trust...")
-        }
-        return true
     }
 }
